Moved userInformation to a new controller.
