+++ MementoMori/MementoMori.Server/Controllers/AuthController.cs
@@ -13,15 +13,13 @@ namespace MementoMori.Server.Controllers
     {
         private readonly IAuthService _authService;
         private readonly IAuthRepo _authRepo;
-        private readonly IDeckHelper _deckHelper;
         private static readonly ConcurrentDictionary<string, User> _registeredUsers = new();
         private static bool initialized = false;
         
-        public AuthController(IAuthService authService, IAuthRepo authRepo, IDeckHelper deckHelper)
+        public AuthController(IAuthService authService, IAuthRepo authRepo)
         {
             _authService = authService;
             _authRepo = authRepo;
-            _deckHelper = deckHelper;
             if (!initialized)
             {
 
@@ -100,28 +98,5 @@ namespace MementoMori.Server.Controllers
             _authService.RemoveCookie(HttpContext);
             return Ok();
         }
-
-        [HttpGet("userInformation")]
-        public async Task<ActionResult> userInformation()
-        {
-            var requesterId = _authService.GetRequesterId(HttpContext);
-            if (requesterId != null)
-            {
-                var userDecks = await _deckHelper.getUserDecks((Guid)requesterId);
-                var userInfo = new UserDeckInformationDTO{
-                    Decks = userDecks ?? [],
-                    IsLoggedIn = true,
-                };
-
-                return Ok(userInfo);
-            }
-            else
-            {
-                var userDecks = new UserDeckInformationDTO{Decks = null, IsLoggedIn = false};
-                return Ok(userDecks);
-            }
-
-        }
-
     }
 }
