+++ MementoMori/mementomori.client/src/DeckCreateButton.tsx
@@ -1,5 +1,3 @@
-/// A placeholder
-import React from 'react';
 import { Button } from '@mui/material';
 const RedirectButton = () => {
   const handleRedirect = () => {
